working on "All" views

+++ ASP.NET Core/PetsDates/PetsDates/Data/Migrations/20210716130928_Pets.cs
@@ -39,7 +39,7 @@ namespace PetsDates.Data.Migrations
                     Id = table.Column<int>(type: "int", nullable: false)
                         .Annotation("SqlServer:Identity", "1, 1"),
                     Name = table.Column<string>(type: "nvarchar(20)", maxLength: 20, nullable: false),
-                    Age = table.Column<int>(type: "int", nullable: false),
+                    Age = table.Column<double>(type: "float", nullable: false),
                     Gender = table.Column<string>(type: "nvarchar(max)", nullable: false),
                     PictureUrl = table.Column<string>(type: "nvarchar(max)", nullable: false),
                     Comment = table.Column<string>(type: "nvarchar(max)", nullable: true),
